+++ controllers/organizerController.js
@@ -1,5 +1,6 @@
 const bcrypt = require("bcrypt");
 const Organizer = require("../models/Organizer.js");
+const mongoose = require("mongoose");
 
 // Criar um organizador
 const createOrganizer = async (req, res) => {
@@ -33,6 +34,14 @@ const getOrganizers = async (req, res) => {
 // Buscar organizador por ID
 const getOrganizerById = async (req, res) => {
   try {
+
+    const { id } = req.params;
+
+    // Verificar se o ID tem o formato correto de ObjectId
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({ error: 'ID inválido' });
+    }
+
     const organizer = await Organizer.findById(req.params.id);
     if (!organizer) {
       return res.status(404).json({ error: "Organizador não encontrado" });
@@ -47,6 +56,14 @@ const getOrganizerById = async (req, res) => {
 // Atualizar um organizador
 const updateOrganizer = async (req, res) => {
   try {
+
+    const { id } = req.params;
+
+    // Verificar se o ID tem o formato correto de ObjectId
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({ error: 'ID inválido' });
+    }
+
     const { name, email } = req.body;
     const organizer = await Organizer.findByIdAndUpdate(req.params.id, { name, email }, { new: true });
     if (!organizer) {
@@ -62,6 +79,14 @@ const updateOrganizer = async (req, res) => {
 // Deletar um organizador
 const deleteOrganizer = async (req, res) => {
   try {
+
+    const { id } = req.params;
+
+    // Verificar se o ID tem o formato correto de ObjectId
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({ error: 'ID inválido' });
+    }
+    
     const organizer = await Organizer.findByIdAndDelete(req.params.id);
     if (!organizer) {
       return res.status(404).json({ error: "Organizador não encontrado" });
